requirements.txt
aiogram==3.5.0
asyncio==3.4.3
python-dotenv==1.0.1
google-generativeai==0.5.3
